+++ .gitignore
@@ -39,3 +39,4 @@ out/
 core/src/main/resources/application.yml
 application-core.yml
 /api/src/main/resources/application.yml
+/core/src/test/resources/application.yml
