+++ api/src/main/java/com/handsup/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.api.config;
+package com.handsup.config;
 
 import static org.springframework.http.HttpHeaders.*;
