+++ api/src/main/java/com/handsup/exception/ErrorResponseTemplate.java
@@ -1,4 +1,4 @@
-package com.api.exception;
+package com.handsup.exception;
 
 public record ErrorResponseTemplate(
 	String message,
