+++ api/src/main/java/com/handsup/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.api.exception;
+package com.handsup.exception;
 
 import java.util.List;
 
@@ -8,8 +8,6 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import com.core.exception.ValidationException;
 
 import lombok.extern.slf4j.Slf4j;
