+++ core/src/main/java/com/handsup/exception/CommonValidationError.java
@@ -1,4 +1,4 @@
-package com.core.exception;
+package com.handsup.exception;
 
 import static lombok.AccessLevel.*;
